+++ Common/Environs.Types.cpp
@@ -1119,7 +1119,7 @@ namespace environs
 
 
 
-		return "Unkown";
+		return "Unknown";
 	} /// -> get()